+++ tmva/tmva/src/BDT/main.cxx
@@ -97,10 +97,10 @@ int main() {
               << std::endl;
   }
 
-  //double event[4] = {6.,148.,72.,35.};
-  std::vector<double> event{1., 115., 70., 30.};
+  //float event[4] = {6.,148.,72.,35.};
+  std::vector<float> event{1., 115., 70., 30.};
   //vector<int> vect{ 10, 20, 30 };
-  //double event[4] = {1.,115.,70.,30.};
+  //float event[4] = {1.,115.,70.,30.};
 
   for (auto& tree : trees){
     //std::cout  << "There are: " << tree.nodes.size() << " nodes\n";
